Move import and export methods to ClipService
