+++ Clipy/Sources/Services/ClipService.swift
@@ -12,6 +12,7 @@ import RealmSwift
 import PINCache
 import RxSwift
 import RxOptional
+import AEXML
 
 final class ClipService {
 
@@ -150,5 +151,43 @@ extension ClipService {
         guard let value = dictionary[type] else { return false }
         guard let number = storeTypes[value] else { return false }
         return number.boolValue
+    }
+}
+
+// MARK: - Import / Expor
+extension ClipService {
+    func exportClipboard() -> AEXMLDocument {
+        let realm = try! Realm()
+        let defaults = UserDefaults.standard
+        let maxHistory = defaults.integer(forKey: Constants.UserDefaults.maxHistorySize)
+        let ascending = !defaults.bool(forKey: Constants.UserDefaults.reorderClipsAfterPasting)
+        let clips = realm.objects(CPYClip.self).sorted(byKeyPath: #keyPath(CPYClip.updateTime), ascending: ascending)
+
+        let xmlDocument = AEXMLDocument()
+        let rootElement = xmlDocument.addChild(name: Constants.HistoryXml.rootElement)
+
+        for (index, clip) in clips.enumerated() {
+            if maxHistory <= index { break }
+
+            guard !clip.isInvalidated else { continue }
+            guard let clipData = NSKeyedUnarchiver.unarchiveObject(withFile: clip.dataPath) as? CPYClipData else { continue }
+            guard !clipData.stringValue.isEmpty else { continue }
+
+            let historyElement = rootElement.addChild(name: Constants.HistoryXml.historyElement)
+            historyElement.addChild(name: Constants.HistoryXml.contentElement, value: clipData.stringValue)
+        }
+
+        return xmlDocument
+    }
+
+    func importClipboard(with xml: AEXMLDocument) {
+        xml[Constants.HistoryXml.rootElement]
+            .children
+            .forEach { historyElement in
+                if let content = historyElement[Constants.HistoryXml.contentElement].value {
+                    let data = CPYClipData(string: content)
+                    save(with: data)
+                }
+            }
     }
 }
